user management and permissions

+++ app/controllers/cats_controller.rb
@@ -11,11 +11,15 @@ class CatsController < ApplicationController
   end
 
   def destroy
-    cat = Cat.find(params[:id])
-    File.delete Rails.root.join('public', 'cats', 'full', cat.id.to_s + cat.extname)
-    File.delete Rails.root.join('public', 'cats', 'thumbnails', cat.id.to_s + cat.extname)
-    cat.destroy
-    redirect_to :action => :index
+    if session[:user] && session[:user][:can_delete_image]
+      cat = Cat.find(params[:id])
+      File.delete Rails.root.join('public', 'cats', 'full', cat.id.to_s + cat.extname)
+      File.delete Rails.root.join('public', 'cats', 'thumbnails', cat.id.to_s + cat.extname)
+      cat.destroy
+      redirect_to :action => :index
+    else
+      render :text => 'permission denied'
+    end
   end
 
   def random
